infrastructure/pyproject.toml
[project]
name = "spring-ai-mcp-agentcore-infra"
version = "0.1.0"
description = "CDK infrastructure for deploying Spring AI MCP server to AgentCore"
requires-python = ">=3.10"
dependencies = [
    "aws-cdk-lib>=2.220.0",
    "constructs>=10.0.79",
    "boto3",
    "awscrt",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
